Added functionality to the button
Now the button prints 'This button works!' when clicked. You can basically do anything. This is just a short example of how to add functionality to the button. Go ahead and test it out. To change its functionality you have rectify the stuff in line 31. Just remove the `print()` statement and do whatever you want.